requirements.txt
certifi==2025.6.15
charset-normalizer==3.4.2
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
pytz==2025.2
requests==2.32.4
smmap==5.0.2
urllib3==2.5.0